Add cursor tracking support to the webserver

+++ src/types.rs
@@ -1,5 +1,5 @@
 pub type Version = logtree::SeqNum;
 pub type Branch = logtree::BranchNum;
 pub type ClientId = u128;
-pub type DocumentPos = usize;
+pub type DocumentPos = u64;
 pub type DocumentId = u128;
